+++ packages/engine/lib/execution/src/package/experiment/comms.rs
@@ -1,3 +1,5 @@
+pub mod update;
+
 use simulation_structure::SimulationShortId;
 
 #[derive(Debug)]
