+++ packages/engine/src/experiment/controller/comms.rs
@@ -92,39 +92,3 @@ pub mod exp_pkg_ctl {
         (ExpPkgCtlSend { inner: send }, ExpPkgCtlRecv { inner: recv })
     }
 }
-
-/// Handles communication between the Experiment Controller and the Experiment Packages for updates
-/// at each simulation step
-pub mod exp_pkg_update {
-    use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender};
-
-    use crate::experiment::{controller::Result, package::StepUpdate};
-
-    #[derive(Clone)]
-    pub struct ExpPkgUpdateSend {
-        inner: UnboundedSender<StepUpdate>,
-    }
-
-    impl ExpPkgUpdateSend {
-        pub async fn send(&self, msg: StepUpdate) -> Result<()> {
-            Ok(self.inner.send(msg)?)
-        }
-    }
-
-    pub struct ExpPkgUpdateRecv {
-        inner: UnboundedReceiver<StepUpdate>,
-    }
-
-    impl ExpPkgUpdateRecv {
-        pub async fn recv(&mut self) -> Option<StepUpdate> {
-            self.inner.recv().await
-        }
-    }
-
-    pub fn new_pair() -> (ExpPkgUpdateSend, ExpPkgUpdateRecv) {
-        let (send, recv) = unbounded_channel();
-        (ExpPkgUpdateSend { inner: send }, ExpPkgUpdateRecv {
-            inner: recv,
-        })
-    }
-}
